fix(notebook): allow automatic package mirrors
The package mirror resolver can select curated regional hosts that the locked Notebook network policy previously blocked. Make the automatic mirror catalog the shared source of truth for both selection and trusted registry hosts, and pin the invariant with a regression test.
